마이페이지 기능 - 프로필 사진 수정 기능 구현, 프로필 사진 외부 경로 핸들러 구현, 외부경로에서 사진 불러오기 구현

+++ src/main/java/com/example/recard/repository/ProfilePhotoRepository.java
@@ -9,13 +9,15 @@ import org.springframework.lang.Nullable;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProfilePhotoRepository extends JpaRepository<ProfilePhoto, Integer> {
+public interface ProfilePhotoRepository extends JpaRepository<ProfilePhoto, Long> {
 
     // p는 profilePhoto 객체 에서 p의 user객체의 userid랑 비교 해서 조회해야함
     // Optional로 jpa가 쿼리를 반환할 때 null값 허용
     // 유저의 프로필 사진을 찾기 위한 쿼리
     @Query("select p from ProfilePhoto p where p.user.user_id = :user_id and deleteAt=null")
     Optional<ProfilePhoto> findByUserId(@Param("user_id") Long user_id);
+
+
